refactor(app): extract Summary, TransactionForm, and TransactionList components
Split monolithic App.jsx into focused child components, each owning
their relevant state (form fields, filters, totals calculation).
