+++ src/TransactionForm.jsx
@@ -1,0 +1,61 @@
+import { useState } from 'react';
+
+const categories = ["food", "housing", "utilities", "transport", "entertainment", "salary", "other"];
+
+function TransactionForm({ onAddTransaction }) {
+  const [description, setDescription] = useState("");
+  const [amount, setAmount] = useState("");
+  const [type, setType] = useState("expense");
+  const [category, setCategory] = useState("food");
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    if (!description || !amount) return;
+
+    onAddTransaction({
+      id: Date.now(),
+      description,
+      amount,
+      type,
+      category,
+      date: new Date().toISOString().split('T')[0],
+    });
+
+    setDescription("");
+    setAmount("");
+    setType("expense");
+    setCategory("food");
+  };
+
+  return (
+    <div className="add-transaction">
+      <h2>Add Transaction</h2>
+      <form onSubmit={handleSubmit}>
+        <input
+          type="text"
+          placeholder="Description"
+          value={description}
+          onChange={(e) => setDescription(e.target.value)}
+        />
+        <input
+          type="number"
+          placeholder="Amount"
+          value={amount}
+          onChange={(e) => setAmount(e.target.value)}
+        />
+        <select value={type} onChange={(e) => setType(e.target.value)}>
+          <option value="income">Income</option>
+          <option value="expense">Expense</option>
+        </select>
+        <select value={category} onChange={(e) => setCategory(e.target.value)}>
+          {categories.map(cat => (
+            <option key={cat} value={cat}>{cat}</option>
+          ))}
+        </select>
+        <button type="submit">Add</button>
+      </form>
+    </div>
+  );
+}
+
+export default TransactionForm;
