+++ src/TransactionList.jsx
@@ -1,0 +1,60 @@
+import { useState } from 'react';
+
+const categories = ["food", "housing", "utilities", "transport", "entertainment", "salary", "other"];
+
+function TransactionList({ transactions }) {
+  const [filterType, setFilterType] = useState("all");
+  const [filterCategory, setFilterCategory] = useState("all");
+
+  let filteredTransactions = transactions;
+  if (filterType !== "all") {
+    filteredTransactions = filteredTransactions.filter(t => t.type === filterType);
+  }
+  if (filterCategory !== "all") {
+    filteredTransactions = filteredTransactions.filter(t => t.category === filterCategory);
+  }
+
+  return (
+    <div className="transactions">
+      <h2>Transactions</h2>
+      <div className="filters">
+        <select value={filterType} onChange={(e) => setFilterType(e.target.value)}>
+          <option value="all">All Types</option>
+          <option value="income">Income</option>
+          <option value="expense">Expense</option>
+        </select>
+        <select value={filterCategory} onChange={(e) => setFilterCategory(e.target.value)}>
+          <option value="all">All Categories</option>
+          {categories.map(cat => (
+            <option key={cat} value={cat}>{cat}</option>
+          ))}
+        </select>
+      </div>
+
+      <table>
+        <thead>
+          <tr>
+            <th>Date</th>
+            <th>Description</th>
+            <th>Category</th>
+            <th>Amount</th>
+          </tr>
+        </thead>
+        <tbody>
+          {filteredTransactions.map(t => (
+            <tr key={t.id}>
+              <td>{t.date}</td>
+              <td>{t.description}</td>
+              <td>{t.category}</td>
+              <td className={t.type === "income" ? "income-amount" : "expense-amount"}>
+                {t.type === "income" ? "+" : "-"}${t.amount}
+              </td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+}
+
+export default TransactionList;
